Move addon_listing_header() (and new_context())
